perf(e2e): split large spec files and increase workers to reduce shard imbalance
Shards 4 and 5 consistently took 30+ minutes while others finished in
5 minutes. Root cause: Playwright shards by file, so the two largest
spec files (pos-checkout 16 tests, products 14 tests) clustered on
the same shards.

Changes:
- Split pos-checkout.spec.ts into pos-cart.spec.ts (cart + menu tests)
  and pos-checkout.spec.ts (order actions + checkout flow)
- Split products.spec.ts into products.spec.ts (POS panel browsing)
  and products-page.spec.ts (Pro/Free drawer page)
- Increase Playwright workers from 2 to 4 (matching ubuntu-latest vCPUs)

14 spec files × 4 projects now distribute more evenly across 6 shards.
Expected improvement: worst shard from ~32 min to ~10 min.
