+++ apps/main/playwright.config.ts
@@ -31,7 +31,7 @@ export default defineConfig<WcposTestOptions>({
 	fullyParallel: true,
 	forbidOnly: !!process.env.CI,
 	retries: process.env.CI ? 2 : 0,
-	workers: process.env.CI ? 2 : 1,
+	workers: process.env.CI ? 4 : 1,
 	reporter: process.env.CI
 		? [
 				['github'],
